refactor: move views, serializers, and files from api/ to the root of each Django app
