+++ acbc_app/academia_blockchain/urls.py
@@ -18,19 +18,16 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.urls import path, include
 
-# from events import views as event_views # TODO
-
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('academia_blockchain.api.urls')),
-    # path('', event_views.events_all, name="events_all"),
     path('accounts/', include('allauth.urls')),
-    path('events/', include('events.urls')),
+    path('api/profiles/', include('profiles.urls', namespace='profiles')),
+    path('api/content/', include('content.urls', namespace='content')),
+    path('api/events/', include('events.urls', namespace='events')),
+    path('ratings/', include('star_ratings.urls', namespace='ratings')),
     # path('exams/', include('exams.urls')),
-    path('profiles/', include('profiles.urls')),
-    path('content/', include('content.urls')),
-    path('ratings/', include('star_ratings.urls', namespace='ratings'))
+    # path('', event_views.events_all, name="events_all"),
 ]
