--- acbc_app/content/api/views.py
@@ -1,7 +1,0 @@
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
-from rest_framework.generics import get_object_or_404
-
-from content.models import Library
-
