+++ acbc_app/content/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
 
 from content.models import Library
-from profiles.api.serializers import UserSerializer
+from profiles.serializers import UserSerializer
